requirements.txt
Flask
rembg
Pillow
requests
